refactor: UserValidator 패키지 수정
